Перенес файлы в соответствии с боевым проектом в директорию main/... из test.
